Revamping the UI code. It still needs some work, but the nice thing is that adding a new screen just means adding one new class (and editing other files to call it, but only one additional class). And most of the non-screen-specific code is in the base class, so there isn't as much copying. This gets rid of all of the event receiver files that need to be in the project as well, so it means that we have fewer files to view.
I also added the first couple of new features to the GUI--
buttons now have a hover state as well as a normal state, and
they can be drawn with a background, but text draw on top (in
fact, this is the default--just tell the button what text to
display, and it will). However, the font selection is a bit
limited (the text is drawn by the irrlicht GUI system, which
uses bitmap fonts, not true type fonts). I put a couple of
new bitmap fonts into the media folder, but we can create
others as well.